refactor(StaticPage): move SideNav out into its own component
