+++ app/components/pages/StaticPages/SideNav.md
@@ -1,0 +1,21 @@
+```js
+<SideNav
+  navList={[
+    {
+      label: 'Contact eLife',
+      link: '/contact-us/contact-elife',
+      component: ContactELife,
+    },
+    {
+      label: 'Editorial Staff',
+      link: '/contact-us/editorial-staff',
+      component: EditorialStaff,
+    },
+    {
+      label: 'Production Staff',
+      link: '/contact-us/production-staff',
+      component: ProductionStaff,
+    },
+  ]}
+/>
+```
